[refactor]-Modularized Back End
