+++ server/models/game_model.js
@@ -1,0 +1,31 @@
+var Sequelize = require('sequelize');
+
+var sequelize = require('../database.js').sequelize;
+
+var Game = sequelize.define('game', {
+  time: {
+    type: Sequelize.DATE,
+    defaultValue: Sequelize.NOW,
+  },
+  token: {
+    type: Sequelize.STRING,
+  },
+  location: {
+    type: Sequelize.STRING,
+  },
+  active: {
+    type: Sequelize.BOOLEAN,
+  },
+});
+
+// var gameModel = {};
+// gameModel.getGame = (requestedToken) => {
+//   Game.findOne({ where:
+//     {
+//       token: requestedToken
+//     }
+//   })
+
+// };
+
+module.exports = Game;
