+++ server/routers/game_router.js
@@ -1,0 +1,6 @@
+const gameController = require('../controllers/game.js');
+const gameRouter = require('express').Router();
+
+gameRouter.get('/', gameController.GET);
+gameRouter.post('/', gameController.POST);
+module.exports = gameRouter;
